refactor(core): move markdown parsing out of the react package
parseInline, parseBlocks and sanitizeHref now live in @mdocui/core and
are exported. They were never React-specific — prose.tsx tokenized the
markdown itself and only used createElement at the end.

@mdocui/react keeps the rendering half and imports the parsing, dropping
108 lines of it. The 32 prose tests pass unchanged, which is the point:
this is a move, not a rewrite.

Adds 14 core tests. sanitizeHref is a security boundary and is now
public API, so it is tested where it lives: javascript:, vbscript:,
data:, file:// and protocol-relative urls all come back undefined.

This is what a non-React renderer needs in order to display prose
without reaching for innerHTML.
