+++ packages/core/src/index.ts
@@ -5,6 +5,8 @@ export { allDefinitions, defaultGroups } from './definitions'
 export type { ParserOptions } from './parser'
 export { StreamingParser } from './parser'
 export { generatePrompt } from './prompt'
+export type { InlineToken, ProseBlock } from './prose'
+export { parseBlocks, parseInline, sanitizeHref } from './prose'
 export type { RegistryOptions } from './registry'
 export { ComponentRegistry, defineComponent } from './registry'
 export type { Token } from './tokenizer'
