+++ packages/react/src/prose.tsx
@@ -1,81 +1,16 @@
+import type { InlineToken } from '@mdocui/core'
+import { parseBlocks, parseInline, sanitizeHref } from '@mdocui/core'
 import { createElement } from 'react'
 
 /**
  * Lightweight built-in markdown renderer for prose nodes.
- * Handles bold, italic, bold+italic, strikethrough, inline code,
- * links, headings (h1-h3), unordered and ordered lists, and
- * paragraph breaks. No external dependencies.
+ *
+ * Parsing lives in @mdocui/core so every renderer shares one implementation;
+ * this file only turns the resulting tokens and blocks into React elements.
  *
  * Consumers who need full GFM support can override via the
  * `renderProse` prop on the Renderer.
  */
-
-interface InlineToken {
-	type: 'text' | 'bold' | 'italic' | 'bolditalic' | 'strikethrough' | 'code' | 'link'
-	content: string
-	href?: string
-}
-
-// Order matters: bold+italic (***) before bold (**) before italic (*)
-const INLINE_PATTERN =
-	/(\*\*\*([^*]+?)\*\*\*|___([^_]+?)___)|(~~([^~]+?)~~)|(\*\*([^*]+?)\*\*|__([^_]+?)__)|(\*([^*\n]+?)\*|_([^_\n]+?)_)|(`([^`]+)`)|(\[([^\][]+)\]\(([^)\s]+)\))/g
-
-function sanitizeHref(href?: string): string | undefined {
-	if (!href) return undefined
-	const trimmed = href.trim()
-	const lower = trimmed.toLowerCase()
-	if (
-		lower.startsWith('http://') ||
-		lower.startsWith('https://') ||
-		lower.startsWith('mailto:') ||
-		lower.startsWith('tel:') ||
-		(trimmed.startsWith('/') && !trimmed.startsWith('//')) ||
-		trimmed.startsWith('#') ||
-		trimmed.startsWith('?')
-	) {
-		return trimmed
-	}
-	return undefined
-}
-
-function parseInline(text: string): InlineToken[] {
-	const tokens: InlineToken[] = []
-	let lastIndex = 0
-
-	for (const match of text.matchAll(INLINE_PATTERN)) {
-		if (match.index > lastIndex) {
-			tokens.push({ type: 'text', content: text.slice(lastIndex, match.index) })
-		}
-
-		if (match[1]) {
-			// Bold+italic: ***text*** or ___text___
-			tokens.push({ type: 'bolditalic', content: match[2] ?? match[3] })
-		} else if (match[4]) {
-			// Strikethrough: ~~text~~
-			tokens.push({ type: 'strikethrough', content: match[5] })
-		} else if (match[6]) {
-			// Bold: **text** or __text__
-			tokens.push({ type: 'bold', content: match[7] ?? match[8] })
-		} else if (match[9]) {
-			// Italic: *text* or _text_
-			tokens.push({ type: 'italic', content: match[10] ?? match[11] })
-		} else if (match[12]) {
-			// Inline code: `code`
-			tokens.push({ type: 'code', content: match[13] })
-		} else if (match[14]) {
-			// Link: [text](url)
-			tokens.push({ type: 'link', content: match[15], href: match[16] })
-		}
-
-		lastIndex = match.index + match[0].length
-	}
-
-	if (lastIndex < text.length) {
-		tokens.push({ type: 'text', content: text.slice(lastIndex) })
-	}
-
-	return tokens
-}
 
 function renderInline(tokens: InlineToken[], keyPrefix: string): React.ReactNode[] {
 	return tokens.map((token, i) => {
@@ -113,114 +48,6 @@ function renderInline(tokens: InlineToken[], keyPrefix: string): React.ReactNode
 				return createElement('span', { key }, token.content)
 		}
 	})
-}
-
-interface Block {
-	type: 'heading' | 'ulist' | 'olist' | 'paragraph'
-	level?: number // for headings: 1-3
-	items?: string[] // for lists
-	content?: string // for heading / paragraph
-}
-
-// Matches markdown headings: 1–3 leading '#' characters, followed by space/tab, then non-whitespace content
-const HEADING_RE = /^(#{1,3})[ \t]+(\S.*)$/
-// Matches unordered list items: leading '-' or '*', followed by space/tab, then non-whitespace content
-const UL_RE = /^[-*][ \t]+(\S.*)$/
-// Matches ordered list items: one or more digits, then '.' or ')', followed by space/tab, then non-whitespace content
-const OL_RE = /^\d+[.)][ \t]+(\S.*)$/
-
-function parseBlocks(content: string): Block[] {
-	const lines = content.split('\n')
-	const blocks: Block[] = []
-	let paraLines: string[] = []
-
-	const flushParagraph = () => {
-		if (paraLines.length > 0) {
-			const text = paraLines.join('\n').trim()
-			if (text) {
-				blocks.push({ type: 'paragraph', content: text })
-			}
-			paraLines = []
-		}
-	}
-
-	let i = 0
-	while (i < lines.length) {
-		const line = lines[i]
-		const trimmed = line.trim()
-
-		// Blank line — flush current paragraph
-		if (!trimmed) {
-			flushParagraph()
-			i++
-			continue
-		}
-
-		// Heading
-		const headingMatch = trimmed.match(HEADING_RE)
-		if (headingMatch) {
-			flushParagraph()
-			blocks.push({
-				type: 'heading',
-				level: headingMatch[1].length,
-				content: headingMatch[2],
-			})
-			i++
-			continue
-		}
-
-		// Unordered list — collect consecutive list items
-		const ulMatch = trimmed.match(UL_RE)
-		if (ulMatch) {
-			flushParagraph()
-			const items: string[] = [ulMatch[1]]
-			i++
-			while (i < lines.length) {
-				const nextTrimmed = lines[i].trim()
-				const nextUl = nextTrimmed.match(UL_RE)
-				if (nextUl) {
-					items.push(nextUl[1])
-					i++
-				} else if (!nextTrimmed) {
-					// blank line ends the list
-					break
-				} else {
-					break
-				}
-			}
-			blocks.push({ type: 'ulist', items })
-			continue
-		}
-
-		// Ordered list — collect consecutive list items
-		const olMatch = trimmed.match(OL_RE)
-		if (olMatch) {
-			flushParagraph()
-			const items: string[] = [olMatch[1]]
-			i++
-			while (i < lines.length) {
-				const nextTrimmed = lines[i].trim()
-				const nextOl = nextTrimmed.match(OL_RE)
-				if (nextOl) {
-					items.push(nextOl[1])
-					i++
-				} else if (!nextTrimmed) {
-					break
-				} else {
-					break
-				}
-			}
-			blocks.push({ type: 'olist', items })
-			continue
-		}
-
-		// Regular text — accumulate into paragraph
-		paraLines.push(line)
-		i++
-	}
-
-	flushParagraph()
-	return blocks
 }
 
 export interface SimpleMarkdownProps {
